tranactiion service errors and clean/build

+++ src/main/java/com/example/bookingmanagementapi/service/TicketAndBookingLogics.java
@@ -17,7 +17,7 @@ public interface TicketAndBookingLogics {
             String alreadyStartedMessage);
 
     BigDecimal getTotalPrice(
-            Integer days,
+            Long days,
             Integer adultNumber,
             List<Integer> childrenAges,
             RoomEntity roomEntity);
